requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
email-validator==2.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
